Drop superfluous CHECK_HEADER_ADD_INCLUDE() call
The include directories are already properly set up for in-tree as well
as phpize builds on Windows.  Checking that timelib_config.h exists
would be nice, but required to add the path for phpize builds, and
since this is not done for other operating systems, it appears to be
overkill for Windows.